chore: added LoadingSpinnerOverlay component
