+++ ts/components/LoadingSpinnerOverlay.tsx
@@ -1,0 +1,78 @@
+import { Button, Text } from "native-base";
+import * as React from "react";
+import { StyleSheet } from "react-native";
+import I18n from "../i18n";
+import variables from "../theme/variables";
+import BoxedRefreshIndicator from "./ui/BoxedRefreshIndicator";
+import { Overlay } from "./ui/Overlay";
+
+const styles = StyleSheet.create({
+  textCaption: {
+    padding: variables.contentPadding
+  },
+  cancelButtonStyle: {
+    alignSelf: "center"
+  }
+});
+
+type Props = Readonly<{
+  isLoading: boolean;
+  loadingCaption?: string;
+  loadingOpacity?: number;
+  onCancel?: any;
+}>;
+
+/**
+ * A HOC to display and overlay spinner conditionally
+ */
+class LoadingSpinnerOverlay extends React.Component<Props> {
+  public render() {
+    const {
+      children,
+      isLoading,
+      loadingCaption,
+      loadingOpacity,
+      onCancel
+    } = this.props;
+    const overlayProps =
+      loadingOpacity !== undefined
+        ? {
+            opacity: 1,
+            backgroundColor: `rgba(255,255,255,${loadingOpacity})`
+          }
+        : undefined;
+    return (
+      <Overlay
+        {...overlayProps}
+        foreground={
+          isLoading ? (
+            <BoxedRefreshIndicator
+              caption={
+                <Text alignCenter={true} style={styles.textCaption}>
+                  {loadingCaption ? loadingCaption : ""}
+                </Text>
+              }
+              action={
+                onCancel && (
+                  <Button
+                    onPress={onCancel}
+                    cancel={true}
+                    style={styles.cancelButtonStyle}
+                  >
+                    <Text>{I18n.t("global.buttons.cancel")}</Text>
+                  </Button>
+                )
+              }
+            />
+          ) : (
+            undefined
+          )
+        }
+      >
+        {children}
+      </Overlay>
+    );
+  }
+}
+
+export default LoadingSpinnerOverlay;
